fix rule ImplementePrivateEndpointApis
